LayoutTests/fast/canvas/webgl/gl-vertexattribpointer.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
  "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<title>WebGL vertexAttribPointer Conformance Tests</title>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="resources/desktop-gl-constants.js" type="text/javascript"></script>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas" width="2" height="2"> </canvas>
<script>
description("This test checks vertexAttribPointer behaviors in WebGL.");

debug("");
debug("Canvas.getContext");

var gl = create3DContext(document.getElementById("canvas"));
if (!gl) {
  testFailed("context does not exist");
} else {
  testPassed("context exists");

  debug("");
  debug("Checking gl.vertexAttribPointer.");

  gl.vertexAttribPointer(0, 3, gl.FLOAT, 0, 0, 12);
  glErrorShouldBe(gl, gl.INVALID_OPERATION,
      "vertexAttribPointer should fail if no buffer is bound");

  var vertexObject = gl.createBuffer();
  gl.bindBuffer(gl.ARRAY_BUFFER, vertexObject);
  gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(512), gl.STATIC_DRAW);
  gl.vertexAttribPointer(0, 1, gl.FLOAT, 0, 256, 0);
  glErrorShouldBe(gl, gl.INVALID_VALUE,
      "WebGL API supports vertex attribute data strides up to 255 bytes");
  gl.vertexAttribPointer(0, 1, gl.FLOAT, 0, 255, 0);
  glErrorShouldBe(gl, gl.NO_ERROR,
      "vertexAttribPointer with stride <= 255 should succeed");
}

debug("");
successfullyParsed = true;

</script>
<script src="../../js/resources/js-test-post.js"></script>

<script>
</script>

</body>
</html>
