LayoutTests/fast/canvas/webgl/gl-vertexattribpointer-expected.txt

This test checks vertexAttribPointer behaviors in WebGL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Canvas.getContext
PASS context exists

Checking gl.vertexAttribPointer.
PASS getError was expected value: INVALID_OPERATION : vertexAttribPointer should fail if no buffer is bound
PASS getError was expected value: INVALID_VALUE : WebGL API supports vertex attribute data strides up to 255 bytes
PASS getError was expected value: NO_ERROR : vertexAttribPointer with stride <= 255 should succeed

PASS successfullyParsed is true

TEST COMPLETE
